// src/app/modules/auth/sign-in/store/models/sign-in.interface.model.ts
/* -------------------------------------------------------------------------- */
//*                          User Data Model Response                          */
/* -------------------------------------------------------------------------- */
export interface UserDataModelResponse {
  code: string | null;
  group: string | null;
  major: string | null;
  password: string | null;
  role: string | null;
  user: string | null;
  username: string | null;
}